backend/requirements.txt
fastapi
uvicorn[standard]
playwright
beautifulsoup4
requests
python-multipart
python-dotenv
google-generativeai
lxml
aiofiles
pydantic
httpx
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
PyJWT
razorpay
